V2.1
- Double critics
- Nerf stone
- Buff ores
- Skill cost adjustments
- Extended material list

+++ src/main/java/com/cludivers/kz2wdprison/beans/ores/OreStats.kt
@@ -3,6 +3,6 @@ package com.cludivers.kz2wdprison.beans.ores
 import jakarta.persistence.Embeddable
 
 @Embeddable
-class OreStats(var maxXp: Int = 0) {
+class OreStats(var maxXp: Float = 0f) {
     var amountMined = 0
 }
